web: fix lint errors by extracting useDocumentTitle hook and moving filter helpers into useMemo
