+++ web/src/components/contributor-detail-page.tsx
@@ -2,7 +2,7 @@ import { useParams, useNavigate } from 'react-router-dom'
 import { useQuery } from '@tanstack/react-query'
 import { Loader2, Building2, AlertCircle, ArrowLeft } from 'lucide-react'
 import { fetchContributor } from '@/lib/api'
-import { useDocumentTitle } from '@/components/page-header'
+import { useDocumentTitle } from '@/hooks/use-document-title'
 
 function formatBps(bps: number): string {
   if (bps === 0) return '—'
