+++ web/src/components/page-header.tsx
@@ -1,15 +1,6 @@
-import { useEffect } from 'react'
 import type { LucideIcon } from 'lucide-react'
 import type { ReactNode } from 'react'
-
-const BASE_TITLE = 'DoubleZero Data'
-
-export function useDocumentTitle(title: string) {
-  useEffect(() => {
-    document.title = `${title} - ${BASE_TITLE}`
-    return () => { document.title = BASE_TITLE }
-  }, [title])
-}
+import { useDocumentTitle } from '@/hooks/use-document-title'
 
 interface PageHeaderProps {
   icon?: LucideIcon
